+++ packages/codegen/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/codegen [1.0.0-alpha.73](https://github.com/anolilab/lunora/compare/%40lunora%2Fcodegen%401.0.0-alpha.72...%40lunora%2Fcodegen%401.0.0-alpha.73) (2026-07-27)
+
+
+### Dependencies
+
+* **@lunora/advisor:** upgraded to 1.0.0-alpha.52
+* **@lunora/agent:** upgraded to 1.0.0-alpha.28
+* **@lunora/server:** upgraded to 1.0.0-alpha.49
+
 ## @lunora/codegen [1.0.0-alpha.72](https://github.com/anolilab/lunora/compare/%40lunora%2Fcodegen%401.0.0-alpha.71...%40lunora%2Fcodegen%401.0.0-alpha.72) (2026-07-27)