+++ packages/testing/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/testing [1.0.0-alpha.74](https://github.com/anolilab/lunora/compare/%40lunora%2Ftesting%401.0.0-alpha.73...%40lunora%2Ftesting%401.0.0-alpha.74) (2026-07-27)
+
+
+### Dependencies
+
+* **@lunora/agent:** upgraded to 1.0.0-alpha.28
+* **@lunora/mail:** upgraded to 1.0.0-alpha.34
+* **@lunora/server:** upgraded to 1.0.0-alpha.49
+
 ## @lunora/testing [1.0.0-alpha.73](https://github.com/anolilab/lunora/compare/%40lunora%2Ftesting%401.0.0-alpha.72...%40lunora%2Ftesting%401.0.0-alpha.73) (2026-07-27)